Moved stuff from conversion/expressions.py and calculation.py into main expressions.py.
